requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
boto3==1.42.2
botocore==1.42.2
cbor2==5.8.0
celery==5.6.0
certifi==2026.1.4
cffi==2.0.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
coverage==7.12.0
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.0.0
decorator==5.2.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.0.14
django-allauth==65.13.1
django-celery-beat==2.8.1
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
django-ratelimit==4.1.0
django-redis==6.0.0
django-silk==5.4.3
django-storages==1.14.6
django-timezone-field==7.1
django-versatileimagefield==3.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
easy-thumbnails==2.10.1
exceptiongroup==1.3.1
executing==2.2.1
factory_boy==3.3.3
Faker==38.2.0
flower==2.0.1
gevent==25.9.1
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gprof2dot==2025.4.14
greenlet==3.3.0
gunicorn==23.0.0
hiredis==3.3.0
humanize==4.14.0
hyperlink==21.0.0
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
matplotlib-inline==0.2.1
moviepy==2.2.1
msgpack==1.1.2
mysqlclient==2.2.7
numpy==2.4.1
oauthlib==3.3.1
packaging==25.0
parso==0.8.5
pillow==10.4.0
pluggy==1.6.0
proglog==0.1.12
prometheus_client==0.23.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-magic-bin==0.4.14
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
qrcode==7.4.2
redis==7.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.4
stack-data==0.6.3
text-unidecode==1.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
zope.event==6.1
zope.interface==8.2
blurhash-python==1.2.2
